pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
rgmining-script = "^0.6.1"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
py-readability-metrics = "^1.4.5"
nltk = "^3.7"
django-crontab = "^0.7.1"
youtube-transcript-api = "^0.4.4"
[tool.poetry.dev-dependencies]
